requirements.txt
beautifulsoup4==4.7.1
celery==4.2.1
Django==2.1.5
django-autofixture==0.12.1
django-celery-beat==1.4.0
django-colorfield==0.1.15
django-cors-headers==2.4.0
django-filter==2.1.0
django-jet==1.0.8
django-polymorphic==2.0.3
django-redis==4.10.0
django-rest-polymorphic==0.1.8
django-map-widgets==0.2.2
django-reversion==3.0.3
django-select2==6.3.1
django-silk==3.0.1
djangorestframework==3.9.1
djangorestframework-camel-case==1.0.3
djangorestframework-gis==0.14
djangorestframework-simplejwt==3.3
drf-flex-fields==0.3.5
drf-yasg==1.14.0
geocoder==1.38.1
Markdown==3.0.1
psycopg2cffi==2.8.1
PyJWT==1.7.1
pyshp==2.1.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
sentry-sdk==0.7.6
six==1.12.0
